✨ Add class sphere
