+++ include/Generation/tools.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <SFML/Graphics.hpp>
+#include <SFML/Window.hpp>
+
+void generateImage(sf::RenderWindow &window, sf::Image &image);
+void displayImage(sf::RenderWindow &window, sf::Image &image);
